refactoring for multi wallet support
