+++ web-ui/src/hooks/useAuth.js
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 
 // auth provider
 //import AuthContext from 'contexts/FirebaseContext';
-import AuthContext from 'contexts/RelayxContext';
+import AuthContext from 'contexts/WalletContext';
 //import AuthContext from 'contexts/TwetchContext';
 // import AuthContext from 'contexts/Auth0Context';
 // import AuthContext from 'contexts/JWTContext';
