+++ web-ui/src/pages/questions/[question_id].js
@@ -49,16 +49,39 @@ const QuestionDetailPage = () => {
       })
       return
     }
-  
-    relayone
-      .send({
-        opReturn: ['onchain', '1HWaEAD5TXC2fWHDiua9Vue3Mf8V1ZmakN', 'answer', json],
-        currency: 'USD',
-        amount: 0.01,
-        to: '1HWaEAD5TXC2fWHDiua9Vue3Mf8V1ZmakN'
+
+    try {
+      switch (wallet){
+        case "relayx":
+          relayone
+            .send({
+              opReturn: ['onchain', '1HWaEAD5TXC2fWHDiua9Vue3Mf8V1ZmakN', 'answer', json],
+              currency: 'USD',
+              amount: 0.01,
+              to: '1HWaEAD5TXC2fWHDiua9Vue3Mf8V1ZmakN'
+            })
+            .then(console.log)
+            .catch(console.error);
+          break;
+        case "twetch":
+          //TODO
+          break;
+        case "handcash":
+          //TODO
+          break;
+        default:
+          console.error("No wallet selected")
+          return
+      }
+    } catch (error) {
+      enqueueSnackbar(`Error Posting Answer: ${error.message}`, {
+        anchorOrigin: {
+          vertical:'top',
+          horizontal:'center'
+        },
+        variant:'error'
       })
-      .then(console.log)
-      .catch(console.error);
+    }
   }
 
   function onAnswer(answer) {
